+++ src/fs/impl/fbdevfs.cr
@@ -48,7 +48,7 @@ private class FbdevFSNode < VFSNode
     state.buffer.to_unsafe.as(UInt8*)
   end
 
-  def ioctl(request : Int32, data : UInt32,
+  def ioctl(request : Int32, data : UInt64,
             process : Multiprocessing::Process? = nil) : Int32
     case request
     when SC_IOCTL_TIOCGWINSZ
